feat: modularize tree-sitter-bash into a separate file (#123)
