+++ src/parsing/mod.rs
@@ -1,3 +1,4 @@
+pub mod bash_parser;
 pub mod cpp_parser;
 pub mod parser;
 pub mod python_parser;
